+++ src/components/decimal/decimal.js
@@ -174,6 +174,10 @@ class Decimal extends React.Component {
     if (this.value === '') {
       this.emitOnChangeCallback('0');
     }
+
+    if (this.props.onBlur) {
+      this.props.onBlur();
+    }
   }
 
   /*