+++ src/components/dropdown-filter-ajax/dropdown-filter-ajax.js
@@ -154,6 +154,10 @@ class DropdownFilterAjax extends DropdownFilter {
       let filter = this.props.create ? this.state.filter : null;
       // close list and reset filter
       this.setState({ open: false, filter: filter });
+
+      if (this.props.onBlur) {
+        this.props.onBlur();
+      }
     }
   }