+++ src/components/dropdown-filter/dropdown-filter.js
@@ -190,6 +190,10 @@ class DropdownFilter extends Dropdown {
           this.emitOnChangeCallback('', this.state.filter);
         }
       }
+
+      if (this.props.onBlur) {
+        this.props.onBlur();
+      }
     }
   }