+++ CHANGELOG.md
@@ -54,6 +54,13 @@ So far we have added attributes to the following components:
 gulp --port 1234
 ```
 
+# 0.34.2
+
+## Bug fix
+
+* Fixes onBlur prop passed to `Date`, `Decimal`, `Dropdown`, `DropdownFilter`, and `DropdownFilterAjax` components so it is called instead of ignored
+* `I18nHelper`: Number abbreviator allows negative numbers
+
 # 0.34.1
 
 ## Component Enhancements